📦 build: switch to pyproject.toml and poetry
Both tsmu.py and transmission-done.py are now console scripts exported
by the package.

Wrapped the contents of transmission-done in a main(), as exported
scripts need a callable as an entrypoint.

Rename transmission-done to transmission-done-standalone, as this
script can be run standalone without additional services. This is in
preparation for the external, service-dependent transmission-done.