✨ feat: add dupes command
The dupes command, when run inside a "dupes" folder as created from the
transmission-done scripts in this project, will check if the dupe is
identical to a parent directory, and if so, re-add it there, and remove
the duplicate copy from the dupes folder.

Initial check in after this script was developed out-of-repo.