🐛 fix: fix finding higher-level parent in dupes
Finding higher-level parents was broken because of a missing break
statement. Without it, we'd always find the last item of
candidate_paths, of which we changed the order in a previous commit.o

Tweak some logging stuff. Don't request 'id' because we don't use it in
an attempt to speed up dupes.